Fix lifecycle crash
The FragmentViewPager and how it interacts with the activity lifecycle
is far from intuitive and heavily misdocumented on the web at this time.

On activity recreate, it creates new fragments for you and initiates their
lifecycle, so creating new instances again when we have saved state leads
to confusion, because these fragments will not be attached or initiated,
and thus likely to not do what you want if you reference some method on them.